Merge in Azure DW code (#1)
* Fixed issue #6.

This issue was caused by the `package_data` variable being setup incorrectly in `setup.py`. Apparently the file listing is **not** recursive, which caused the `sdist` files to lack several macro overrides needed to function.

* Added a note about the ODBC driver

As noted in issue #5, the use of the `driver` variable was not particularly clear. I've added some info to try and explain that configuration better.

* altered mssql__create_view_as macros

Looks like this issue is being caused by the `mssql__create_view_as` macro. It works fine if the sql it's provided does not contain a CTE. If it does, sql server considers it a syntax error.

This commit removes the parenthesis wrapping the `{{ sql }}` portion.

* Fixed CTEs with insert into statements

SQL Server's `insert into` syntax isn't nearly as forgiving as in other databases. In the previous version I had created a cte as apart of the into statement that could later be referenced.

This worked so long as your source model didn't contain a CTE of its own. If it did, that put a CTE declaration inside another CTE which broke everything. I've taken a que from the dbt-sqlserver package and am now creating a "temporary" view to handle the issue.

Thanks @mikaelene for the example.

* incremented version number

* Update README.md

* updated .gitignore

* delt with empty column names

It appears that issue #10 is caused by MSSQL not returning default column names for aggregate functions through the odbc library. When the `''` column name hits the agate library, an error is thrown.

To handle this behavior, I've overridden the class method, `get_result_from_cursor()`. The new method loops through all the column names and replaces any instances of `''` with `unnamed_column-{i}`. This should provide a simple work around that the user doesn't really see, but is also very easy for the user to avoid if it's undesired behaivor.

* incremented version

* Update README.md

* initial commit

* added catalog

* added work from https://github.com/norton120/dbt-azuredatawarehouse

* moved to explicit varchar size per https://docs.microsoft.com/en-us/azure/sql-data-warehouse/sql-data-warehouse-tables-data-types

* handle empty string column name from scalar

* update README to match dockerfile

* added sample profile.yml

* clean up git conflict noise in readme

* known issues in readme

* added process_results class method and updated sql connection type

Co-authored-by: Jacob Mastel <jacobm001@users.noreply.github.com>
Co-authored-by: Isaac Chavez <ichavez@1aauto.com>
Co-authored-by: Ethan Knox <norton120@gmail.com>